# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recordatorios API",
    "version": "0.2.0"
  },
  "paths": {
    "/api/v1/reminders/preview": {
      "post": {
        "summary": "Preview Route",
        "operationId": "preview_route_api_v1_reminders_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePreview"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reminders": {
      "post": {
        "summary": "Create Route",
        "operationId": "create_route_api_v1_reminders_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPublic"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Route",
        "operationId": "list_route_api_v1_reminders_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "upcoming",
                "fired",
                "canceled"
              ],
              "type": "string",
              "default": "upcoming",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPage"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reminders/{reminder_id}/cancel": {
      "post": {
        "summary": "Cancel Route",
        "operationId": "cancel_route_api_v1_reminders__reminder_id__cancel_post",
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reminder Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderPublic"
                }
              }
            }
          },
          "401": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "type": {
                      "title": "Type",
                      "type": "string"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "traceId": {
                      "title": "Traceid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "traceId"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReminderCreate": {
        "properties": {
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "local_time": {
            "type": "string",
            "title": "Local Time"
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone"
          },
          "message": {
            "type": "string",
            "maxLength": 280,
            "minLength": 1,
            "title": "Message"
          },
          "send_whatsapp": {
            "type": "boolean",
            "title": "Send Whatsapp",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "local_date",
          "local_time",
          "timezone",
          "message"
        ],
        "title": "ReminderCreate"
      },
      "ReminderPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReminderPublic"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ReminderPage"
      },
...
      "ScheduleInput": {
        "properties": {
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "local_time": {
            "type": "string",
            "title": "Local Time"
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "local_date",
          "local_time",
          "timezone"
        ],
        "title": "ScheduleInput"
      },
      "SchedulePreview": {
        "properties": {
          "scheduled_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At Utc"
          },
          "local_date": {
            "type": "string",
            "title": "Local Date"
          },
          "local_time": {
            "type": "string",
            "title": "Local Time"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "resolution": {
            "type": "string",
            "enum": [
              "exact",
              "gap_forward",
              "overlap_later"
            ],
            "title": "Resolution"
          }
        },
        "type": "object",
        "required": [
          "scheduled_at_utc",
          "local_date",
          "local_time",
          "timezone",
          "resolution"
        ],
        "title": "SchedulePreview"
      },
...
      "VersionInput": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "VersionInput"
      }
    }
  }
}